修复了若干问题

+++ jjj_shop_multi_shop/src/components/file/Upload.vue
@@ -377,7 +377,15 @@
                 temp_list.push(item.fileId);
               }
             }
-          }
+          };
+          if(temp_list.length==0){
+              loading.close();
+              self.$message({
+                  message: '请选择文件',
+                  type: 'warning'
+              });
+              return ;
+          };
           FileApi.deleteFiles({
             fileIds:temp_list.join(',')
           }, true)
